Function hyperscan_sys::hs_scan_vector

pub unsafe extern "C" fn hs_scan_vector(
    db: *const hs_database_t, 
    data: *const *const c_char, 
    length: *const c_uint, 
    count: c_uint, 
    flags: c_uint, 
    scratch: *mut hs_scratch_t, 
    onEvent: match_event_handler, 
    context: *mut c_void
) -> hs_error_t

The vectored regular expression scanner.

This is the function call in which the actual pattern matching takes place for vectoring-mode pattern databases.

@param db A compiled pattern database.

@param data An array of pointers to the data blocks to be scanned.

@param length An array of lengths (in bytes) of each data block to scan.

@param count Number of data blocks to scan. This should correspond to the size of of the @p data and @p length arrays.

@param flags Flags modifying the behaviour of this function. This parameter is provided for future use and is unused at present.

@param scratch A per-thread scratch space allocated by @ref hs_alloc_scratch() for this database.

@param onEvent Pointer to a match event callback function. If a NULL pointer is given, no matches will be returned.

@param context The user defined pointer which will be passed to the callback function.

@return Returns @ref HS_SUCCESS on success; @ref HS_SCAN_TERMINATED if the match callback indicated that scanning should stop; other values on error.